cache-hotness: replace modulus with bitwise and
As access_per_cache_line grows, the effect is quite sizeable. On my
machine I ran a few tests with vs without the optimization:

With optimization:

    $ usr/local/bin/cache-hotness -m 64m -a 1 -y 512 -c no
    Execution time parent: 5.206156731 s
    $ usr/local/bin/cache-hotness -m 64m -a 7 -y 512 -c no
    Execution time parent: 6.075818578 s
    $ usr/local/bin/cache-hotness -m 64m -a 53 -y 512 -c no
    Execution time parent: 17.467309092 s

Without optimization:

    $ usr/local/bin/cache-hotness -m 64m -a 1 -y 512 -c no
    Execution time parent: 5.122098078 s
    $ usr/local/bin/cache-hotness -m 64m -a 7 -y 512 -c no
    Execution time parent: 12.230176243 s
    $ usr/local/bin/cache-hotness -m 64m -a 53 -y 512 -c no
    Execution time parent: 89.880489695 s

When access_per_cache_line is 1, we are clearly stalled by memory
accesses. But, as access_per_cache_line grows, the role of the actual
computation on the CPU grows. When running with -a 53, the running time
is 17.5s vs 89.9s, or a 414% increase.